requirements.txt
pygame-ce
moderngl
scikit-learn
joblib